warning
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@10935 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ plugins/vocaliser/vocaliser.cpp
@@ -291,7 +291,7 @@ bool PlaysndCommand::handle ( int playerID, bzApiString _command, bzApiString _m
 		if (item.team)
 			target = playerInfo->team;
 
-		voice.lastVoiceTime = time;
+		voice.lastVoiceTime = (int)time;
 
 		bz_sendTextMessage(playerID,target,item.text.c_str());
 		bz_sendPlayCustomLocalSound (target, item.sound.c_str());
